ditch bad settings panel
