+++ resources/editor/imgui_config.ini
@@ -54,6 +54,11 @@ Size=1050,56
 Collapsed=0
 DockId=0x00000009,0
 
+[Window][Images]
+Pos=161,117
+Size=937,548
+Collapsed=0
+
 [Docking][Data]
 DockSpace         ID=0x11111111 Window=0x1BBC0F80 Pos=0,22 Size=1600,878 Split=X Selected=0x8065DB7E
   DockNode        ID=0x00000001 Parent=0x11111111 SizeRef=251,701 Selected=0x698DF469
